requirement.txt
Distance==0.1.3
nltk==3.4.5
Pillow==6.1.0
pkg-resources==0.0.0
six==1.12.0
tqdm==4.36.1